+++ client/src/components/myprofile/mystation/MyStationItem.js
@@ -1,10 +1,15 @@
-const MyStationItem = ({ stamp }) => {
+import { useRecoilValue } from "recoil";
+import { trainInfo } from "../../../atoms/trainInfo";
+
+const MyStationItem = ({ trainId }) => {
+  const trainInfoList = useRecoilValue(trainInfo);
+  console.log(trainInfoList[trainId - 1].train, trainId);
   return (
     <div className="">
       <img
-        className="m-4 w-20 h-20 rounded-md hover-img"
-        alt="post img"
-        src={stamp}
+        className="m-4 w-20 h-20 rounded-md hover:scale-150" //hover-img
+        alt="postImg"
+        src={`/images/기차역도장/${trainInfoList[trainId - 1].train}.png`}
       />
     </div>
   );
